#27: o.l.c.m.NotBlank matcher
